Added Fishing areas along with Fishing animation and movement functionality
Fishing mechanic and animations added, can move freely between fishing and moving. Done in a fairly hacky way  but it's fine to refactor later.

To add: particle effect of Nitin's fish coming up whenever fishing animation loops; add 1 to a score meter whenever this happens; create a fishing bar to show the amount of time it will take to gain a new fish.